+++ src/awlists/awlists.go
@@ -28,7 +28,7 @@ func (l *WList) Allocate(vertices int) {
  * checking if that neighbor was already inserted
  * in the given vertex and weigth.
  */
-func (l *WList) Relate(vertex, neighbor int, weigth float32, edges *int) {
+func (l *WList) Relate(vertex, neighbor int, weigth float64, edges *int) {
 	v := vertex - 1
 	n := neighbor - 1
